Add GH::Events::Template::UnknownDictFileError
Its better to have a CLI class that handles interfacing with the
terminal, so that we don't have to deal with system exits in specs and
library usage.

This replaces the system exit by raising an error and handling it in the
CLI callsite.

It would be better to handle this bit of logic in a CLI class, but this
is an improvement for now I think.