Extract GH::Events::Text class
Rather than using a module method, this commit refactors
`GH::Events::Text` in to a class.

The logic in `.translate` is only subtly different, this will allow us
to use the class in more interesting ways, without worrying too much
about the inputs.

This refactoring makes it easier to test the basic logic of the
rendering once – in the instance methods – and add specific tests for
new class methods that massage the inputs in to the correct form.