Extract GH::Parser::WithType class
GH::Events::Text.translate still does way too much work.

This commit extracts the "type" parsing in to the `WithType` class.

The class is a `SimpleDelegator` [1], which delegates most calls to the
encapsulated `Hash`, but adds the extra attributes that we need.

Its a bit hacky to parse the JSON initially, then convert back to JSON
to parse into an `OpenStruct` later, but I think the reduction of
complexity is worth it for now. Simply passing the `Hash` to
`OpenStruct` doesn't give us the deep method chaining that parsing JSON
into an `OpenStruct` does:

    h = { foo: { bar: 'baz' } }

    o = OpenStruct.new(h)
    # => #<OpenStruct foo={:bar=>"baz"}>
    o.foo
    # => {:bar=>"baz"}
    o.foo.bar
    # => NoMethodError: undefined method `bar' for {:bar=>"baz"}:Hash

    o = JSON.parse(h.to_json, object_class: OpenStruct)
    # => #<OpenStruct foo=#<OpenStruct bar="baz">>
    o.foo
    # => #<OpenStruct bar="baz">
    o.foo.bar
    # => "baz"

We could parse in to an `OpenStruct` initially, adding the new
attributes with setters, but the delegation behaviour of
`SimpleDelegator` doesn't work well with `OpenStruct` when the API of
the object differs due to differing data in the JSON. Here's an example:

    os = OpenStruct.new(foo: 'bar')
    # => #<OpenStruct foo="bar">
    os.respond_to?(:baz)
    # => false
    os.baz
    # => nil

    class Delegated < SimpleDelegator ; end

    d = Delegated.new(os)
    d.respond_to?(:baz)
    # => false
    d.baz
    # => NoMethodError (undefined method `baz' for #<OpenStruct … >)

So, for now I think its cleaner to just munge all our extra attributes
on to a `Hash` and then parse our final `Hash` in to an `OpenStruct`.

[1] https://ruby-doc.org/stdlib-2.6.3/libdoc/delegate/rdoc/SimpleDelegator.html