Extract GH::Parser::WithRef class
Extract the final bits of complexity from `GH::Events::Text`.

This uses the same pattern as `WithType` – simply adding extra
attributes to our parsed `Hash` before we convert it to an `OpenStruct`.

There's a little unneeded complexity here as we're overwriting the `ref`
attribute. I didn't want to break the API at this point, but in the
future we could keep `ref` as the full ref (e.g. `"refs/heads/foo"`) and
add an extra `short_ref` attribute.